refactor(settlement): move the settlement API functions to their own module
