+++ crates/interledger-settlement/Cargo.toml
@@ -11,7 +11,7 @@ repository = "https://github.com/interledger-rs/interledger-rs"
 bytes = { version = "0.4.12", default-features = false }
 futures = { version = "0.1.29", default-features = false }
 hyper = { version = "0.12.35", default-features = false }
-interledger-http = { path = "../interledger-http", version = "^0.3.0", default-features = false, features = ["idempotency"] }
+interledger-http = { path = "../interledger-http", version = "^0.3.0", default-features = false }
 interledger-packet = { path = "../interledger-packet", version = "^0.3.0", default-features = false }
 interledger-service = { path = "../interledger-service", version = "^0.3.0", default-features = false }
 log = { version = "0.4.8", default-features = false }
@@ -27,6 +27,7 @@ tokio = { version = "0.1.22", default-features = false }
 num-bigint = { version = "0.2.3", default-features = false, features = ["std"] }
 num-traits = { version = "0.2.8", default-features = false }
 warp = { version = "0.1.20", default-features = false }
+http = "0.1.19"
 
 [dev-dependencies]
 parking_lot = { version = "0.9.0", default-features = false }
