+++ crates/interledger-settlement/src/api/mod.rs
@@ -1,0 +1,13 @@
+mod api;
+mod client;
+mod message_service;
+
+#[cfg(test)]
+mod fixtures;
+#[cfg(test)]
+mod test_helpers;
+
+// Expose the API creation filter method and the necessary services
+pub use api::create_settlements_filter;
+pub use client::SettlementClient;
+pub use message_service::SettlementMessageService;
